+++ crates/yast-core/src/lua/widgets/internal.rs
@@ -1,13 +1,55 @@
+use log::{debug, error, info, trace, warn};
 use mlua::prelude::*;
 
 use crate::lua::widgets::LuaWidget;
 
 pub(super) fn init_internals(lua: &Lua) -> LuaResult<()> {
-  let table = lua.create_table()?;
+  let children_table = lua.create_table()?;
 
   let unique = lua.create_function(|_, index: usize| Ok(LuaWidget::InternalChild(index - 1)))?;
-  table.set("get", unique)?;
+  children_table.set("get", unique)?;
 
-  lua.globals().set("children", table)?;
+  lua.globals().set("children", children_table)?;
+
+  let logging_table = lua.create_table()?;
+
+  logging_table.set(
+    "debug",
+    lua.create_function(|_, msg: String| {
+      debug!("lualog: {}", msg);
+      Ok(())
+    })?,
+  )?;
+  logging_table.set(
+    "trace",
+    lua.create_function(|_, msg: String| {
+      trace!("lualog: {}", msg);
+      Ok(())
+    })?,
+  )?;
+  logging_table.set(
+    "info",
+    lua.create_function(|_, msg: String| {
+      info!("lualog: {}", msg);
+      Ok(())
+    })?,
+  )?;
+  logging_table.set(
+    "warn",
+    lua.create_function(|_, msg: String| {
+      warn!("lualog: {}", msg);
+      Ok(())
+    })?,
+  )?;
+  logging_table.set(
+    "error",
+    lua.create_function(|_, msg: String| {
+      error!("lualog: {}", msg);
+      Ok(())
+    })?,
+  )?;
+
+  lua.globals().set("log", logging_table)?;
+
   Ok(())
 }